requirements.txt
requests
beautifulsoup4
streamlit
playwright
openai
faiss-cpu
numpy
python-dotenv
sqlalchemy
alembic